requirements.txt
pymilvus==2.4.0
timm==1.0.7
torch==2.2.2
Pillow==10.3.0
scikit-learn==1.5.0
Numpy==1.24.1
ipython==8.13.0
